+++ NEWS.md
@@ -6,6 +6,9 @@ dplyr 0.1.2
 * The hybrid evaluation did not handle some expressions correctly, for example in
   `if( n() > 5 ) 1 else 2` the subexpression `n()` was not substituted correctly. 
 
+* `row_number` gives correct results (#227). 
+
+* `filter` now fails when given anything that does not evaluates to a logical vector
 
 dplyr 0.1.1
 -----------